Update all existing controllers with reality of authentication

+++ backend/src/main/java/com/example/service/WarehouseService.java
@@ -1,6 +1,5 @@
 package com.example.service;
 
-import com.example.dto.PostWarehouseRequestDto;
 import com.example.dto.PostWarehouseResponseDto;
 import com.example.dto.WarehouseDto;
 import com.example.model.User;
@@ -23,7 +22,14 @@ public class WarehouseService {
         this.userRepository = userRepository;
     }
 
-    public PostWarehouseResponseDto createWarehouse(Integer userId, String name, String address) {
+    public List<WarehouseDto> getAllUsersWarehouses(int id) {
+        return userRepository.findById(id)
+                .map(User::getWarehouseList)
+                .map(wl -> wl.stream().map(this::convertToWarehouseDto).collect(Collectors.toList()))
+                .orElseThrow();
+    }
+
+    public PostWarehouseResponseDto createWarehouse(int userId, String name, String address) {
         Warehouse warehouse = new Warehouse();
         warehouse.setUser(userRepository.findById(userId).orElseThrow());
         warehouse.setName(name);
@@ -31,23 +37,22 @@ public class WarehouseService {
         return convertToPostWarehouseResponseDto(warehouseRepository.save(warehouse));
     }
 
-    public List<WarehouseDto> getAllUsersWarehouses(Integer id) {
-        return userRepository.findById(id)
-                .map(User::getWarehouseList)
-                .map(wl -> wl.stream().map(this::convertToWarehouseDto).collect(Collectors.toList()))
-                .orElseThrow();
-    }
-    //fix it redundant userID
-    public void deleteWarehouse(Integer userId, Integer warehouseId) {
+    public PostWarehouseResponseDto updateWarehouse(int userId, int warehouseId, String name, String address) {
         Warehouse warehouse = warehouseRepository.findById(warehouseId).orElseThrow();
-        warehouseRepository.delete(warehouse);
-    }
-    //fix it redundant userID
-    public PostWarehouseResponseDto updateWarehouse(Integer userId, Integer warehouseId, String name, String address) {
-        Warehouse warehouse = warehouseRepository.findById(warehouseId).orElseThrow();
+        if (warehouse.getUser().getId() != userId) {
+            throw new RuntimeException(userId + " cannot update this warehouse");
+        }
         warehouse.setName(name);
         warehouse.setAddress(address);
         return convertToPostWarehouseResponseDto(warehouseRepository.save(warehouse));
+    }
+
+    public void deleteWarehouse(int userId, int warehouseId) {
+        Warehouse warehouse = warehouseRepository.findById(warehouseId).orElseThrow();
+        if (warehouse.getUser().getId() != userId) {
+            throw new RuntimeException(userId + " cannot delete this warehouse");
+        }
+        warehouseRepository.delete(warehouse);
     }
 
     private PostWarehouseResponseDto convertToPostWarehouseResponseDto(Warehouse warehouse) {
@@ -56,7 +61,6 @@ public class WarehouseService {
         postWarehouseResponseDto.setAddress(warehouse.getAddress());
         return postWarehouseResponseDto;
     }
-
 
     private WarehouseDto convertToWarehouseDto(Warehouse warehouse) {
         WarehouseDto warehouseDto = new WarehouseDto();
